requirements.txt
streamlit==1.9.2
pandas==1.4.2
base58==2.1.0
streamlit-option-menu==0.3.2
pillow==9.1.1
scipy==1.8.1
seaborn==0.11.2
matplotlib==3.5.2